Refactor into functions to make program flow clear.
All the work to determine colours of the cells, the links to the individual graphs, onclick listeners etc are all very similar and can be extracted out into functions. Various variables that are reused often can also be given more meaningful names that describe what they contain (rather than random offsets). This makes it much clearer what is going on in the main body of the code.