Remove unnecessary code when building tracemaps.
It doesn't appear to require copies of various structures, so the tracemap work doesn't need to be duplicated.

The REST API will now also return standard deviations, so most of the work on the rtt data is not needed (and some of it was extra, duplicate work that was not required).