Second phase of the URL changeover
Muninbytes now works with the new URL structure. Completely new dropdowns for
the munin page -- dropdowns for each of switch, interface and direction.

The various updateDropdown functions are now named better: initXDropdown().
The purpose has also changed: they figure out which (if any) dropdown values
are selected to ensure the internal variables are set properly and ensure the
dropdown lists are sorted when first loading the page.

Updated basicts.js to use Brendon's new event styling. The script also has been
updated to ensure it sets the URL correctly, including all the timestamps.

Disabled a whole bunch of useless URL-related code from changeGraph().

Changed smokeping data queries in the API code to match recent changes in ampy.

Fixed introduced bug whereby the dropdowns were not set to have appropriate
values for the current stream, especially when the stream was manually specified
in the URL.

Fixed another introduced bug where manually setting the timestamps in the URL
(or presumably just refreshing an existing page) would cause the wrong
timestamps to be used to draw the initial graph.