Switched ordering of timestamps in URLs
Timestamps in URLs are now detailstart/detailend/summarystart/summaryend.

This is to support creating URLs based on events -- we know when the event
occurred so can calculate the range for the detailed graph but want to use
sensible defaults for the summary. With this change, we can leave off the
summary timestamps and (hopefully) get ampweb to decide what the summary
should show.

However, at the moment our default is to just show the last month of data,
regardless of where the detailed view is looking. We will try to fix this
soon.

Also removed a bunch of unused or commented-out code from nntscgraph.js.
Documented the various new functions a little better too, so the script file
should be a lot nicer to work with.