Fixed broken titles on graph pages
Titles will now appear on pages for NNTSC-based graphs. The title is based off
the 'name' field for the stream in NNTSC.

Added a new api call: _streaminfo. This returns all of the information
available about a given stream id as a dictionary. It is primarily used for
getting the name string for the page title, but could be useful in other ways.

The URL is now updated after every call to changeGraph rather than requiring a
separate call to updatePageURL.