Fixed incorrect dropdown values for some streams
When there are two streams that populate a dropdown box with similar strings,
the incorrect string was often being designated as selected when populating
the dropdown box. This was because we were deciding who was selected by using
object:contains() which does not require an exact match -- if the text for
the option "contains" the string you asked for, then it will return true.

This meant that a stream with a destination of "google.com" was always setting
"google.com.v6" as the selected option, because the latter string contained
the string we were looking for.

Now we make sure we do an exact match when we have a known selected value.