Better documentation of the contains bug fix
Made sure I documented why I was not using object.contains() in the code
itself, in case someone comes along later and decides to be "smart".

Optimised the code somewhat as well. We now break out of the loop function once
we've found the text value we were looking for.

If there is no selected value, we now use the same code (looking for the magic
--SELECT-- option) rather than a second branch that still used
object.contains().