Improved summary graph x-axis
Calculate sensible default values for the start and end times of the summary
graph based on the selected region.

Replaced useless automagic xtic labels in the summary graph with a set of
labels that are generated programatically by us. Now we can have ticks on
every day, but only put labels on a limited number of days to avoid problems
with labels overlapping one another. Might need to be reduced to less than
15 labels per graph, though....