Removed hard-coded NNTSC and netevmon parameters
The location of both the NNTSC exporter and the netevmon database are now
specified in the development.ini file, rather than being hard-coded throughout
our source files. This makes it much easier for anyone to change this
configuration.

Note: this solution is rather specific to the pyramid / pylons setup we are
using for development and testing. Once we move to a proper apache setup, we'll
need to move this config somewhere else.