Removed summary graph timestamps from URL
The boundaries of the summary graph are no longer displayed in the URL, nor can
they be explicitly set via the URL.

Instead, I've inserted a new "summary scale" value into the URL between the
stream id and the detailed timestamps. This describes how many days worth of
data should be used as a baseline for the automagic generation of the summary
graph.

This should mean that refreshing a graph page should update the summary with
any new data that has appeared since the last refresh. It also means that we
can more easily implement zooming in and out on the summary graph and it
tidies up the URL somewhat.

I've also removed the pointless code in graph.py to generate the initial
timestamps -- these values were in the wrong order but aren't used anymore
anyway, so I've gotten rid of it all to remove any potential confusion.