Attempted fix for off-by-one error when selecting on summary graph.
This appears to fix the problem we are seeing where the right edge of the selected area is truncated, so that the resulting detail graph does not show the full area. Who knows what side effects this has, but the function appears to be only used for selecting on graphs.

This issue appears to be similar to what we are seeing:

    https://github.com/HumbleSoftware/envisionjs/issues/21

Keep track of what is going on with this in the trac, Ticket #60.