Moved History init into a common function
Brad had complained that there was duplicated code for initialising history.js,
so I've created a new script file (util.js) for storing 'common' or duplicated
code. The first entry in this file is a startHistory function which should
solve Brad's complaint and hopefully we can move other little bits and pieces
of code into there as well.