Fix the grid being drawn at the wrong scale after extending the Y axis.
The initial grid is based on the datapoints in the series, ignoring the values of the smoke. The Y axis is later extended to cover all the smoke values but the grid was not being updated, possibly resulting in a very tightly packed grid if the axis was massively extended. It should now recalculate the ticksize based on the new axis maximum.