Fix bad xlabel on summary graph in IE 10
Two more ways in which javascript is stupid:
 * You can't just ask for the timezone of the client -- you have to use a dirty
   hack based on parsing the result of Date.toString()
 * There's no standard way of implementing Date.toString(), so of course IE 10
   gives you an entirely different string to everyone else

Added a little function to util.js for getting a suitable timezone string. It's
a very dirty hack but manages to do something useful for IE 10 as well as
real web browsers. Still need to make sure it works for old IEs, as well as
figure out what Opera does.