Replaced dropdown code with OO implementation
The dropdown list code was getting unmanagable so I've replaced it with a
much nicer object-oriented implementation. Now it should be easier to add
dropdown lists for new collections -- just implement a new Dropdown subclass
with all the required methods and make sure the appropriate .js file is
included by the graph view.

Much better than having to edit a crapload of switch statements all over the
place to make sure the right dropdown function is called for the collection
you are trying to add.