Merge the mousewheel/mousedrag scrolling functions. Fix edge cases.
The mousewheel and mouse dragging functions were doing almost the same work and didn't need to be duplicated. They now use the same function with just the delta calculated differently per event type.

Also tidied up the cases where the selected area hit the edge of the summary graph to make sure that it will go all the way to the edge, and once there no spurious requests are made.