Added graphs for lpi-packets and lpi-flows
These graphs all use the LPIBasic code path, so were pretty easy to implement.
The only slightly tricky part is lpi-flows being split into two metrics in
the dropdown list: one for peak flows and one for new flows. We end up tacking
a little submetric parameter on the end of our _streams request based on the
metric selected in the dropdown list.