Added graphs for LPI user counts
This required adding a new dropdown class and page template. We do the same
thing with switching between 'active' and 'observed' user streams as we did
with the 'peak' and 'new' flows last time around.

It's currently not possible to switch metrics from a "user" style metric to a
"basic" metric or vice versa, but switching between metrics of the same style
(e.g. bytes to packets or flows) seems to work fine.