Use a single template for all of the graph pages
Instead of having a template per collection, where the only real difference
was the explicit definition of each dropdown on that page, we now have a
single 'graph' template which uses TAL to automatically place dropdowns
according to a list of dropdowns passed in via the same python that was
populating the dropdowns anyway.

Updated the CSS to make sure the dropdowns look nice and are placed reasonably
sensibly, e.g. no dropdowns to the right of the graph, nice padding between
dropdowns.

The old 'graph.pt' template has been renamed 'ampgraph.pt' as it has the old
sparklines code in it which we'll want at some future point. graph.pt now
contains the new template added in this commit.

Updated the dropdown javascript code to no longer use object.name in their
callbacks, as I realised we could just use object.id instead and save ourselves
having to specify an otherwise unused field for each dropdown list.

Removed the old per-collection templates as they are no longer needed.