Improved user selection on LPI graphs
The user dropdown is now dependent on the source, protocol and direction
dropdowns -- this means the dropdown will only show users that have existing
streams for that source / protocol / direction combo. Previously, the user
list was simply populated with all users that had been observed for the
source.

However, switching the protocol or direction while looking at a graph for a
particular stream doesn't necessarily invalidate the user selection. If the
user also has a stream for the new protocol / source / direction combo, the
user will stay selected and the graph will automatically switch to the new
stream. If the user doesn't have a stream, the selection is changed to
"--SELECT--" and the graph doesn't change until a new valid user is chosen.

The user dropdown has also moved to be last in the 'list', since it depends
on everything else now.