Moved collection-specific code out of api.py and graph.py
Each collection should now define a sub-class of CollectionGraph that
implements all of the methods. These methods perform all of the collection
specific stuff that was cluttering up api.py and graph.py.

This will make it easier to add new collections to amp-web, as well as
keep the main code-base more manageable and easy to understand.