Completely reworked nntscgraph.js
nntscgraph.js was a hideous mass of code so I've split it into separate
components.

graph.js is the top-level stuff:
  * the "on-page-load" functionality
  * responding to history events, e.g. the user pushing "Back"
  * managing the Graph object for the page
  * providing an API for other scripts to interact with the Graph object

betternntscgraph.js defines NNTSCGraph, the superclass that all Graph objects
are derived from. All of the code common to all NNTSC graph objects is included
in here as functions in the NNTSCGraph class. This includes functions for
updating the page URL, generate X tics, changing the zoom level, etc. The bulk
of the code from nntscgraph.js has ended up in here.

Finally, I've created a graphobjects directly which contains a set of .js
files: one for each collection. Each defines a subclass of NNTSCGraph which
implements a couple of methods that are collection-specific -- drawing the
graph itself and calling the right dropdown constructor.

Adding a new collection should be much easier now -- add a .js file to
graphobjects/ implementing a subclass of NNTSCGraph and update
createGraphObject in graph.js to call the constructor for the new
Graph object you implemented.