Started to bring back tabs showing related graphs alongside the main graph
This commit is phase one. The new ampy API for getting related streams is
used to find out what tabs we should be drawing and we go ahead and draw them.

We're mostly using the old CSS styling from the student's implementation. I've
rescued the styling that was hidden away in old javascript for the selected
tab and created a class for it instead. I've also tweaked the font for the
tab label to make it take up less space in the tab and look nicer in general.
More fiddling with the styling will occur in due course.

Clicking on the tabs doesn't do anything yet and the preview graphs are
missing, but these will be targets for subsequent phases.