Started placing dropdowns inside javascript
Prior to this, our dropdowns were being specified and placed using a
combination of Python and TAL. This meant that we had to do a full page
reload if the collection changed (and therefore the current dropdowns were
rendered invalid).

I'm now moving towards getting a list of dropdowns via an ajax call and
letting javascript clear and populate the dropdown div as required. We
can re-use the populateDropdown function for much of the work, but I did
change it so that it can populate disabled dropdowns without re-enabling them
by default. It is now the dropdown callback's responsibility to re-enable
dropdowns.

I've made the new dropdown approach work for the two amp collections. Will
continue migrating the other collections over tomorrow.

Also fixed a couple of bugs in the AMP dropdown code -- silly stuff like
missing '>'s on the end of elements.