Limit the possible range of bin sizes
Previously, we tried to work out the optimal binsize given the range of our
graph. This makes it very hard to cache data usefully, as every individual
zoom level is a different binsize and therefore can't be cached.

Now we work out what the optimal size would be and then round up to the next
10 minute boundary (aside from special cases at very high zoom levels), which
means that you can zoom in or out a few 'clicks' and not need to fetch new
aggregated data from the database.