Fix issue with the y-axis not aligning with the detail data
The maximum value for the y-axis in the detail graph was being determined
across all of the available data, including the summary data that was not
being rendered in the graph. Now we run a quick check over the actual data
that will be displayed and force the maximum y axis to a sensible value based
on that.

Ensure that the y-axis is also updated sensibly while the user is in the
process of selecting an area to look at using the summary graph.

Added our own version of the selection plugin for flotr2. This will form the
basis of a few modifications to the standard behaviour which we don't really
want. So far, I've already changed it so that just clicking on the summary
graph is a null operation rather than clearing the selection. The way I've
done this will result in a new query for the same data as is being shown
already, which isn't ideal but avoids the problem of the handles disappearing
(which is what occurs if you just break out of the mouseup callback early).