If we get a new request for tabs, cancel any ongoing ones
This fixes the bug observed for Chrome where the tabs would be duplicated when
a page is first loaded.
This is due to Chrome firing a popstate event on pageLoad, which triggers an
additional call to changeStream. This in turn causes the tabs to be regenerated
(because we've supposedly changed stream).