Moving towards showing a useful date at high zoom levels
Started adding our own tickFormatter function for the detail graph so we can
try to insert date information when the zoom level is high. At this stage, it
only partly works -- the date is shown on the midnight tick only rather than
on the first visible tic for that day.

Tic formatting when zoomed out still works the same as it did before, which
is an achievement in itself since we've had to replicate the default flotr
behaviour in our new function.

Also removed an old foo() function that never should have been committed in
the first place :/