Show the date as well as time at high zoom levels
Detail graphs that cover a range of 36 hours or less now include the date on
the xtic labels. This is applied to all xtics (rather than selectively, as
with an earlier commit) and looks a bit crappy.

Ideally, we'd be able to put the date on its own separate line but flotr likes
to replace newline characters with spaces so this is a lot harder than it
sounds.