Implemented auto-scaling of the summary graph
Also, completely re-designed the architecture of the static/scripts directory.
Apologies for doing this all in one big commit -- everything was a bit too
intertwined to split up nicely.

Autoscaling
-----------
The range of the summary graph is now automatically determined based on the
size of the selected area. As soon as you select more than 75% of the current
summary, the summary range will be extended. If you make a subsequent smaller
selection (less than 75% of the next smallest possible summary range), the
summary shrinks again.

This enables us to get rid of the zoom buttons on the graph. Also, the
right side of the summary graph is no longer fixed to now, so exploring data
older than a month won't always result in a huge summary range.

There are still a few features needed:
 * auto-panning the summary when you pan the detail graph to the edge of the
   summary range
 * preventing the left edge of the summary range from going past the start of
   the stream (i.e. into nothingness).

basicts.js rewrite
------------------
To facilitate the autoscaling summary graph, some major changes had to be
made to basicts.js. Previously, this code only allowed for changes to be made
to the detail graph without re-creating the entire set of graphs. The querying
approach was also inefficient and the code style was unpleasant.

I've rewritten the whole lot to be more object-oriented and better able to
support both present and potential future needs. You now instantiate an
instance of the BasicTimeSeriesGraph class and call methods such as
drawGraphs and updateSummaryGraph to trigger changes in the displayed graph
components. The AJAX queries for data are now triggered simultaneously rather
than waiting for the first request to finish before making the second etc.

The BasicTimeSeriesGraph class can also be subclassed to implement different
graph styles by overriding a handful of methods. In this way, I've implemented
a SmokepingGraph class inside smoke.js, which has its own methods for
determining the Y axis range and setting the Flotr line configuration.

I've also split some of the contents of the BasicTimeSeriesGraph class out
into separate files. config.js contains all of the default Flotr configuration
for the graphs. ticlabels.js contains the code for generating nice X axis tic
labels for both graphs. interaction.js deals with creating the envision
components and defining callbacks for various interactions with the graphs,
e.g. zooming, panning, selection.

Javascript restructure
----------------------
There were a few problems with the existing code structure:
 * Too many .js files inside of static/scripts
 * Vague or confusing terminology: graphObjects, NNTSCGraph, templates
 * A lot of external code mixed in with scripts that we are actively developing
 * Some script files are still too long and difficult to work in

I've made a few changes to try and fix some of these problems.

1. "Graph Page" is the term I'm trying to use to describe the webpage that
   has the graphs on it. It includes the dropdowns and tabs. Everything that
   was previously a Graph Object should now be known as a Graph Page.
2. NNTSCGraph has become a CuzGraphPage. betternntscgraph.js has been renamed
   cuzgraphpage.js.
3. Graph templates have become "Graph Styles". There are only two styles at
   the moment: BasicTimeSeriesGraph and SmokepingGraph. Styles describe how
   to fetch data for, draw and interact with both the detail and summary
   graphs.
4. Scripts that weren't written by us have been moved into a lib/ directory
   so they don't clutter up the directories where we are actively working.
5. Script inclusion is all done within views/graph.py. There are no
   collection-specific script includes now, as we are going to need them all
   anyway to move between collections without triggering page-loads.
6. Deleted some old scripts that aren't being used anymore.