No longer tracking start and end times in cuzgraphpage.js
The CuzGraphPage class used to do a lot of the manipulation of start and end
times for the summary and detail graphs. However, this can all be done inside
basicts.js.

This means that we don't need to track start and end times in multiple places
and removes a lot of code from cuzgraphpage.js. Quite a bit of code has moved
into graph.js instead, e.g. code for managing the URL.

The drawGraph method in graphPage objects now takes a start and end time
which defines the initial selection for the new graph. Updated all of the
graphPage classes to accept these parameters correctly.