Avoid showing useless whitespace in the summary graph
We now cap the left edge of the summary graph to avoid showing large amounts
of "empty space", i.e. time before the stream first started collecting data.

This is done by requesting the timestamp of the first datapoint for the stream
from NNTSC and using that as a limit upon the left edge of the summary graph
(plus a little padding to make it obvious there is no more data to the left).
This does add an extra API request, though, which we must wait for before
actually creating the graphs themselves.

If the first data timestamp is not easily available (e.g. in streams created
before we added this feature to NNTSC), we can use the first datapoint in the
summary data as an indication. This works well in most cases, but will have
problems if there is a large gap in the data at the place where the left edge
of the summary graph would fall.

However, we always ensure that at least 1 week is shown in the summary
graph regardless of how much data is available. This is to avoid the problem
of a 1 hour summary graph when a stream is first created.

Made a couple of tweaks to the generateSummaryXTics function to account for
the possibility of a summary graph being smaller than 30 days in length.