Added support for multiple stream ids in the URL
Additional stream ids in the URL may now be specified by separating them with
hyphens ('-'). This is a first step towards supporting drawing multiple streams
on the same graph.

At the moment, doing this will simply result in the first stream being drawn
on the graph but internally all of the specified streams are stored as an
array (previously only a single numeric id was expected). Outwardly, everything
should work the same as before, except multiple streams should be able to be
specified in the URL without breaking everything.

Changed the webpage title behaviour to be more sensible in the presence of
multiple streams. If only one stream is selected, the title is the name string
from the streaminfo as before. If multiple streams are specified, a more
generic title based on the collection is used.