Ensure all selected streams are passed into the graph drawing code
This also means I should be able to avoid having to make too many changes
to the collection-specific code in future.

The base URLs for graph-related queries no longer include the stream id. This
is now added by basicts.js just prior to making the query, so it will in
theory be able to easily construct queries for multiple streams in future.

Fixed key errors in the dropdown saving and restoring code.