Stream list now contains objects rather than just ids
This will allow us to store more information with the stream that will help
with plotting -- line colour, show/hide, group. At the moment, id is the
only attribute stored in the stream object but it will be much easier to
add other attributes now.

Clicking on a tab now calls a new function called changeTab rather than
changeGraph. This is because the dropdowns only provide a list of stream ids
rather than stream objects, so changeGraph will convert them into objects.
By contrast, changeTab takes a list of objects as the tab generation code
aims to ensure that related streams have the same properties as the matching
streams shown on the current graph and thus generates the new stream objects
itself.

Ultimately, the dropdowns will go away soon so it wasn't worth the effort of
updating all the dropdown code to pass a list of objects into the dropdown
callback instead of the list of ids.

Fixed bug where multiple tabs would be selected if they matched the same
collection (e.g. new flows and peak flows both came from lpi-flows so both
would be highlighted if either was selected).