Enabled "address" dropdowns for amp-icmp and amp-traceroute graphs
AmpTracerouteGraph class in views/collections has now become a sub-class of
AmpIcmpGraph due to the amount of common code.

There are some issues with the dropdown selections shown when moving through
history -- in particular, the 'address' is not set correctly if the destination
changes -- but the dropdowns will be going away very soon so I've not
bothered to fix it for now.