First go at modal interface for dealing with views.
Currently works only with amp-icmp, replacing the dropdown system that was in place. Still needs to be tidied up so that javascript functions appear in and are called from sensible places (rather than being in the global scope), and there is lots of crufty code that will no longer be required and can be removed.