Added event markers for the summary graph.
Added a blank title to the summary graph to pad the top offset slightly as is done with the detail graph, which revealed a bug in the handles plugin not taking that offset into account when positioning drag handles. Included a fixed version of the plugin (handles.js) to temporarily address this issue.