Added a new eventview/ URL which creates a view from a event stream id
Unlike streamview/ which will only show that stream explicitly, eventview
will construct a group that includes all streams that share similar properties
to the stream id given in the event.

This is because event detection (for some collections) looks at an aggregated
group of streams, rather than a single stream, but is identified using only
one of the streams in that group. The line we show on the graph therefore also
needs to represent that group.