Fixed event tooltips to be less debuggy and more useful
The event tooltips that show up when you hover over an event
on the graph pages have been updated to better describe the line
that the event belongs to. Now when there are multiple lines being
drawn, it should be easier to tell which line the event belongs to.

Also removed the date and timezone portions of the timestamp shown
in the event tooltips as this is not useful in most cases.

Made the severity score more obvious. Display the event metric only
for the collections where there are multiple possibilities.