Add a basic modal template for the rainbow graph
Note: the modal doesn't actually work at all -- waiting until
the modal branch is finished so it can be implemented the "right"
way.

Made sure the right modal is displayed for the rainbow graph. This
required adding a new member variable to each GraphPage, as we
now have this distinction between the collection where the data
comes from and the type of graph used to display the data, i.e.
it isn't a one-to-one mapping anymore.