Fixed rainbow graph to show detailed data properly
Previously, the rainbow style only generated suitable points for the
summary data which meant that the heavily aggregated data would appear
in the detail graph as well.

In the process, moved the code to generate the detail data series by merging
it into the longer summary series to a separate function so that it could
be used by the rainbow style as well.