Group consecutive identical hops into a single hop entry
This fixes the problem where a path with multiple consecutive
0.0.0.0's would not form a single nice box when moused over
in the rainbow graph.

Set the maximum Y axis value to 4 hops in the rainbow graph in cases
where there are less than four hops. Otherwise this will be set to the
largest hopcount + 1. This change makes the rainbow graph less in
your face when there is only one hop and leaves a bit of white space for
drawing event markers.