Add ability to aggregate ipv4 or ipv6 only to amp-icmp modal.
This means it is now possible to get an ipv4 aggregated line without
having to choose ipv4/ipv6 split and end up with a ipv6 line
cluttering things up :)

amp-traceroute also gains this feature by virtue of inheritance.

Note: the button names are a bit ugly but we might be able to tidy
this up once we add tooltips for radio buttons.