Added very basic Graph Browser page.
The page simply allows a user to select a collection and be taken
directly to the view page for that collection.

I've put absolutely no effort into making it look nice. This commit is
purely to provide the functionality.

Every collection must now provide a get_browser_collections function which
returns a list of dictionaries describing each graphable collection.

Moved some utility code from view.py into a new common.py. This code deals
with common tasks like connecting to ampy and generating the correct
instance of a collection class for a given collection name.