Combine events for the same group into a single marker on a graph
"Severity" on the tooltip has been replaced with the number of
detectors that reported the event (i.e. the group size). The colour
of the marker is still based on severity (which will eventually be
a useful value that represents the severity of the event group rather
than individual events).