Fix bug with clicking on traceroute tabs
There was one last check for the old 'modifier' field that I had
removed recently.

Replaced 'modifier' in all the tab definitions with 'graphstyle'. We still
need something to say "draw this traceroute data using a rainbow graph instead
of a line graph" unfortunately. By defining the graphstyle separate from the
data collection we can do this, albeit in a somewhat clunky way -- for most
tabs the collection is the same as the graphstyle, for example.