Removed need for distinction between graphstyle and collection
Tabs now simply describe the graph style they will be showing and
the data collection is inferred automatically using a new
graphStyleToCollection function.

Made sure eventview, streamview, tabview and validatetab all consistently
use graphstyle in their URL and convert that to collection using the new
function.

To aid this, event links for amp-traceroute streams now use
amp-traceroute-rainbow rather than amp-traceroute.

Every collection view class now provides a get_event_graphstyle function
which returns the graphstyle to place in an event link for that collection.
By default, this is simply the collection name so it only needs to overridden
in cases where this is not the case (e.g. the amp-traceroute case above).