Added a new flotr type for drawing basic time series lines
This allows us to colour the lines for BasicTS graphs in the same
way that we colour smokeping and the legend.

Moved the colour calculation code into util.js so that we are only doing
it in one place, as opposed to the three places we would need to otherwise
(smokeping, basicts and when drawing the legend).