Fixed bug where rainbow graph summarys were incorrect.
The main problem was that we weren't setting the maximum y axis value
for the summary graph at all, which meant the rainbow drawing code was
was unable to place the yaxis points sensibly.

Amazingly we had gotten away with not doing this up until now, presumably
because envision was able to autoscale conventional time series graphs.
The rainbow graph was just too different for it to deal with.

Also removed some debug output that had crept into develop.