Made event labels on the dashboard more concise
Removed information that is now duplicated in the description
provided by netevmon.

Also removed stream detail that is not overly relevant to a person
considering clicking on the event, e.g. packet sizes, destination
IP addresses, etc.

There may still be some value in including information about the
detector that caught the event somewhere (in a tooltip perhaps?),
mainly as this will help distinguish between TEntropy Stddev and
TEntropy Propdiff events which otherwise look exactly the same
and occur close to one another. Otherwise these may be interpreted
as duplicates or an escalation by users.