Display graphs as the ajax completes rather than waiting for everything
This means we can display the detail graph as soon as data is available
for it rather than waiting for the summary data (which can take a long
time if it isn't cached!).

Still a few cosmetic issues to fix -- mainly removing all traces of
the old loading graphic -- but seems to be functional.