Refactor matrix popovers so code can be reused in ajax error handlers.
Much of this code was going to be repeated in displaying both good and bad popovers, so I pulled it out into functions. While I was there I also moved a whole lot of code, that though it was only called once, really made the main logic hard to follow. It should now hopefully be clearer what is actually happening at each step when trying to display a popover.