Fixed problems with colour inconsistencies on graphs
This includes problems such as:
 * colours fluctuating as the graph was loading
 * line colours not matching the legend

We now use the colour ids provided in the legend data to colour
both the lines and the legend, so we're finally consistent rather
than relying on the data being in a certain sorted order.

This has meant the data format for graphs has changed. Instead of
'data' being just a sorted array of datapoints for plotting, it is
now an object with a 'colourid' and a 'series' (where series is
the old sorted array). This was the only sensible way for me to
get the colourid into the plot functions.

Hopefully I've updated all the graph types to use this new format
properly...