Try to minimise sluggishness due to summary loading
The main cause of the sluggishness is the time taken to redraw the summary
graph when new summary data arrives. This is mostly a problem on voodoo, due
to shitty remote X -- tests on standalone machines seem to have much
less problems but it's still a bit annoying.

Reduced the number of points shown on the summary graph (which in turn
increases the binsize). Less points == less time spent drawing them. This
reduction will also apply to the detail graph if it covers more than a few
days.

Only redraw the summary graph when we have at least a week of undrawn data.
Special cases are the first and last fetch, which will always trigger a
redraw.