Lower recent duration for traceroute data to 10 minutes to match icmp.
This was previously at a longer duration because the test was only performed every 10 minutes, so we needed to guarantee that we would have data to display. It is now performed every minute, so we should have plenty of data to use (and matching duration makes the standalone caching script easier).