Separate out Flotr2 from Envision and update Flotr2 to latest version.
Envision depends on parts of Flotr2 but hasn't been updated in 2 years, while Flotr2 receives regular updates. This allows us to remove the grid plugin we had, as the same version is included in the full version of Flotr2. Dependencies of both Flotr2 and Envision (bean and underscore) have also been moved into separate files.

Also removed the canvas2png script as it was never used, and would have only been usable with a PHP backend anyway.

Renamed files for consistency with .min where the contents have been minified.