Handle query timeouts when loading the matrix
This also fixes the empty day_data bug that Brad has been complaining
about.

Raise a suitable error message so the user knows why their matrix is
empty.

Don't fetch day_data for the absolute latency matrix, as we don't
actually need it. Recent data is much quicker to fetch so we should
timeout much less often on this matrix.