Fix summary graph bug where the line would jump backwards
The most important part of the fix is that we now check any
summary data we receive and remove any datapoints that are for
timestamps that we've already got from a previous fetch. This
prevents us creating overlaps in our summary data series.

We now update 'fetched' to be the earliest timestamp in our
summary series rather than being set to the earliest timestamp
we asked for. Due to block alignment, we often get a few extra
measurements either side so now 'fetched' better reflects what we
have available.

Also added a distinction between fetching data for the detail graph
("full") and the summary graph ("summary"), so ampy can avoid
sending us extra blocks for the summary graph that we are just going
to discard anyway. Updated all collections to not include "full" in
their data request URL automatically.