Move "Recursion" ahead of "Query" in DNS modal
This matches the recent ampy change where the recursion field has
been moved much earlier in the stream property hierarchy.

Also replaced the various updateX functions with a generic function
that forms the correct _destinations URL and runs the AJAX query for
the next selectable. This removes a lot of silly copy-pasted code
from the modal javascript and makes it much easier to modify the
existing modals.