Extend dns matrix cell to last 30 mins
Some DNS tests have a frequency > 10 mins, so colouring our
cells based on the last 10 mins of data tends to result in
a lot of pink squares.

Ideally, the tooltip would also adjust itself to not show
the last 10 mins as well if no data is available, but that's
less important.