Merged amp-icmp and amp-dns into a single graph page
Now we can plot both DNS and ICMP test results on the same graph.

Created a new modal for choosing series that uses a tabbable
area to switch between ICMP and DNS (and their respective
specific options).

Still a lot of legacy ICMP/DNS specific code to be removed at
some point -- traceroute tends to inherit from a lot of the
ICMP stuff so will need to be careful about what we keep and
what we remove.