Fix disabled tab href being added to URL
Clicking on a disabled tab would cause the href (e.g.
'#dns' being added to the current URL, which is both ugly and
prevents people from being able to refresh the page easily (as
the modified URL is no longer valid).

Solution is not the nicest -- add special onclick events for every
tab that check if the tab is disabled and prevents default behaviour
if so.

Also added a space to "AMP TCP Ping" label on the graph browser for
readability.